kiosk/requirements.txt
opencv-python>=4.8.0
pyzbar>=0.1.9
requests>=2.31.0
python-dotenv>=1.0.0
